Fixed implicit conversion issue of mdev calculation

+++ ping.h
@@ -192,8 +192,8 @@ extern jmp_buf pr_addr_jmp;
 extern int timing;			/* flag to do timing */
 extern long tmin;			/* minimum round trip time */
 extern long tmax;			/* maximum round trip time */
-extern long long tsum;			/* sum of all times, for doing average */
-extern long long tsum2;
+extern double tsum;			/* sum of all times, for doing average */
+extern double tsum2;
 extern int rtt;
 extern __u16 acked;
 extern int pipesize;
